refactor: split chart component
